♻️ Separate auth dto logic
