+++ server/src/auth/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './create-user.dto';
+export * from './login.dto';
